Refactor prepareLinks into utils/prepareLinks.js
